Inline visitor, remove code duplication in loader.cc in favor of templates / std::enable_if
